+++ aleo/artifacts/js/leo2js/wusdc_token.ts
@@ -9,7 +9,7 @@ import * as leo2js from "./common";
 export function gettoken(token: tokenLeo): token {
   const result: token = {
     owner: leo2js.address(token.owner),
-    amount: leo2js.u64(token.amount),
+    amount: leo2js.u128(token.amount),
     _nonce: leo2js.group(token._nonce),
   }
   return result;
